docs: remove SECURITY.md
We have an org-wide SECURITY.md file that will be shown by GitHub
in the repository UI automatically. This points the users to
security@grafana.com and the bug bounty programme.

I assume it is a relic of the past that we still reference
security@pyroscope.io here; we do not believe we check this regularly in
the security team.